Adding leaderboard modal
