+++ packages/web/src/pages/asset-detail-page/components/AssetLeaderboardCard.tsx
@@ -1,7 +1,24 @@
 import { useArtistCoinMembers, useUsers } from '@audius/common/api'
-import { Flex, Paper, Text, Divider, Skeleton } from '@audius/harmony'
+import {
+  Flex,
+  Paper,
+  Text,
+  Divider,
+  Skeleton,
+  IconCaretRight,
+  IconButton
+} from '@audius/harmony'
+import { useDispatch } from 'react-redux'
 
 import { UserProfilePictureList } from 'components/notification/Notification/components/UserProfilePictureList'
+import {
+  setUsers,
+  setVisibility
+} from 'store/application/ui/userListModal/slice'
+import {
+  UserListEntityType,
+  UserListType
+} from 'store/application/ui/userListModal/types'
 
 import { AssetDetailProps } from '../types'
 
@@ -16,8 +33,19 @@ export const AssetLeaderboardCard = ({ mint }: AssetDetailProps) => {
   const { data: users, isPending: isUsersPending } = useUsers(
     leaderboardUsers?.map((user) => user.user_id)
   )
-
+  const dispatch = useDispatch()
   const isPending = isLeaderboardPending || isUsersPending
+
+  const handleViewLeaderboard = () => {
+    dispatch(
+      setUsers({
+        userListType: UserListType.COIN_LEADERBOARD,
+        entityType: UserListEntityType.USER,
+        entity: mint
+      })
+    )
+    dispatch(setVisibility(true))
+  }
 
   return (
     <Paper
@@ -32,7 +60,13 @@ export const AssetLeaderboardCard = ({ mint }: AssetDetailProps) => {
         </Text>
       </Flex>
       <Divider css={{ width: '100%' }} />
-      <Flex pv='l' ph='xl'>
+      <Flex
+        pv='l'
+        ph='xl'
+        justifyContent='space-between'
+        alignItems='center'
+        w='100%'
+      >
         {isPending ? (
           <Skeleton />
         ) : (
@@ -45,6 +79,12 @@ export const AssetLeaderboardCard = ({ mint }: AssetDetailProps) => {
             stopPropagation={true}
           />
         )}
+        <IconButton
+          color='subdued'
+          icon={IconCaretRight}
+          aria-label='Open the leaderboard modal'
+          onClick={handleViewLeaderboard}
+        />
       </Flex>
     </Paper>
   )
